requirements.txt
flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
jinja2==3.1.5
